+++ CHANGELOG.md
@@ -10,6 +10,7 @@ INTRO
   - The game now begins in Verdin Village.
   - Time is ticking as soon as you start.
   - Most cutscenes in Adon have been either completely changed or removed.
+- Increased max name length from 10 → 12.
 - Reduced drop rate of Ring of the Woods/Forest.
 
 ### Bug Fixes